Move couch design service to libs
