+++ src/services/warm-views.ts
@@ -3,7 +3,7 @@ import * as Context from 'effect/Context';
 import { Array } from 'effect';
 import { getDbNames } from '../libs/couch/dbs-info';
 import { CouchDesignDocsService } from './couch/design-docs';
-import { getViewNames } from './couch/design';
+import { getViewNames } from '../libs/couch/design';
 import { CouchViewService } from './couch/view';
 import { CouchDesignInfoService } from './couch/design-info';
 import { ChtClientService } from './cht-client';
